[OKD-Core] Check namespace before deleting it
We check if the namespace (project) exists before the infrastructure
tries to delete it. This is important because some tests (like the
metalkube tests) don't create a namespace and don't include
base.scenario.ts in the beginning of the suite - so the infrastructure
shouldn't fail in the end when it tries to delete it.