Monitoring: Fix unmount / mount of whole page when navigating
`PollerPages` was sometimes created by the `Route` in `MonitoringUI` and
sometimes by the `Route` in `AlertingPage`. This meant that it was
being unmounted and mounted again when navigating around the Monitoring
UI. This caused unnecessary rendering and unnecessary calls to the
Prometheus API.

This fixes that by only creating `PollerPages` from `MonitoringUI`.