Query Browser: Fix series toggle buttons when queries are disabled
Don't discard empty query results because that causes the list of all
query results and the list of all disabled data series to go out of
sync.

This was causing the data series (lines in the graph) toggle buttons to
break if an earlier query returned no data (e.g. if it was disabled).