Fix react warning when open the project overview screen without 'Project Access' permissions.
When open the screen without this permissions, `false` was added to the
list of pages. In the `DetailsPage` > `HorizontalNav` this pages was
used to render a list of `Route`s where the key was then `(false).name`

https://github.com/openshift/console/blob/master/frontend/public/components/utils/horizontal-nav.tsx#L239

Warning: Each child in a list should have a unique "key" prop.
See https://fb.me/react-warning-keys for more information.