Allow cluster admins to create terminals in the openshift-terminal namespace
Modify backend proxy check to allow cluster admins if and only if their
terminal is in the openshift-terminal namespace

Only search for terminals in openshift-terminal namespace when using cluster admin

Rework terminal proxy to use isClusterAdmin for checking if the user can access the terminal

Use error status to see if namespace exists in CloudShellAdminSetup

Return false if namespace is not found in CloudShellAdminSetup instead of displaying error.
The reason we need to do this is if you set the error the terminal will flash quickly with
the error before switching to the newly created terminal.

Make subject access review request against openshift-terminal namespace

Display loading screen until admin check has finished

Signed-off-by: Josh Pinkney <joshpinkney@gmail.com>