Disallow executable permissions for all groups.
Reduce permissions from 755 to 644 thus not allowing any group to
directly execute this test (say, in a Node.js environment that had
Cypress APIs injected). This is unarguably a hypothetical case,
possible, but not probable, that something like this would happen.
Nonetheless, this PR promotes uniformity among test cases and their
respective permissions.

Signed-off-by: Pranshu Srivastava <rexagod@gmail.com>